Renamed `SpotlessExtesionBase` to `SpotlessExtension`, and the implementation details live in `SpotlessExtensionImpl`.
