+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessPlugin.java
@@ -35,7 +35,7 @@ public class SpotlessPlugin implements Plugin<Project> {
 		project.getPlugins().apply(BasePlugin.class);
 
 		// setup the extension
-		project.getExtensions().create(SpotlessExtension.EXTENSION, SpotlessExtension.class, project);
+		project.getExtensions().create(SpotlessExtension.class, SpotlessExtension.EXTENSION, SpotlessExtensionImpl.class, project);
 
 		// clear spotless' cache when the user does a clean
 		project.getTasks().named(BasePlugin.CLEAN_TASK_NAME).configure(clean -> {
